[frontend] integrate Convex auth provider
