+++ frontend/src/app/ClientLayout.tsx
@@ -1,0 +1,37 @@
+'use client';
+
+import { Inter } from "next/font/google";
+import "./globals.css";
+import { ThemeProvider } from "@/components/theme-provider";
+import { Toaster } from "@/components/ui/toaster";
+import { ConvexClientProvider } from "./ConvexClientProvider";
+import { cn } from "@/lib/utils";
+import 'katex/dist/katex.min.css';
+
+const inter = Inter({ subsets: ["latin"] });
+
+export default function ClientLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en" suppressHydrationWarning>
+      <body className={cn(inter.className, "h-screen w-screen flex flex-col overflow-hidden bg-background")} suppressHydrationWarning>
+        <ConvexClientProvider>
+          <ThemeProvider
+            attribute="class"
+            defaultTheme="system"
+            enableSystem
+            disableTransitionOnChange
+          >
+            <main className="flex-1 overflow-auto">
+              {children}
+            </main>
+            <Toaster />
+          </ThemeProvider>
+        </ConvexClientProvider>
+      </body>
+    </html>
+  );
+}
