+++ frontend/src/app/layout.tsx
@@ -1,37 +1,17 @@
-'use client';
+import { metadata } from './metadata';
+import ClientLayout from './ClientLayout';
+import { ConvexAuthNextjsServerProvider } from "@convex-dev/auth/nextjs/server";
 
-import { Inter } from "next/font/google";
-import "./globals.css";
-import { ThemeProvider } from "@/components/theme-provider";
-import { Toaster } from "@/components/ui/toaster";
-import { ConvexClientProvider } from "./ConvexClientProvider";
-import { cn } from "@/lib/utils";
-import 'katex/dist/katex.min.css';
+export { metadata };
 
-const inter = Inter({ subsets: ["latin"] });
-
-export default function ClientLayout({
+export default function RootLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode;
-}>) {
+}) {
   return (
-    <html lang="en" suppressHydrationWarning>
-      <body className={cn(inter.className, "h-screen w-screen flex flex-col overflow-hidden bg-background")} suppressHydrationWarning>
-        <ConvexClientProvider>
-          <ThemeProvider
-            attribute="class"
-            defaultTheme="system"
-            enableSystem
-            disableTransitionOnChange
-          >
-            <main className="flex-1 overflow-auto">
-              {children}
-            </main>
-            <Toaster />
-          </ThemeProvider>
-        </ConvexClientProvider>
-      </body>
-    </html>
+    <ConvexAuthNextjsServerProvider>
+      <ClientLayout>{children}</ClientLayout>
+    </ConvexAuthNextjsServerProvider>
   );
 }
